# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "名单管理系统",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/name-list",
    "paths": {
...
        "/namespace": {
            "get": {
                "description": "命名空间查找功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命名空间"
                ],
                "summary": "命名空间查找功能",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code 命名空间code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PageIndex 页码",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PageSize 分页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortedField 排序字段",
                        "name": "sortedField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复查询成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResultPaged"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Namespace"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "命名空间创建功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命名空间"
                ],
                "summary": "命名空间创建功能",
                "parameters": [
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/namespace.CreateNamespaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复创建成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Namespace"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/namespace/{id}": {
            "delete": {
                "description": "命名空间删除功能, 通过namespace_id删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命名空间"
                ],
                "summary": "命名空间删除功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "命名空间id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResult"
                        }
                    }
                }
            },
            "patch": {
                "description": "命名空间修改功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命名空间"
                ],
                "summary": "命名空间修改功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "命名空间id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/namespace.UpdateNamespaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复更新成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Namespace"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/type/{id}": {
            "delete": {
                "description": "名单类型删除功能, 通过list_type_id删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型删除功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复删除成功的名单类型数据",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResult"
                        }
                    }
                }
            },
            "patch": {
                "description": "名单类型修改功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型修改功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list_type.UpdateListTypeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复更新成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListType"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpx.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回包code，表明是否正确，在code == 0时，表明服务正常",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "message": {
                    "description": "回报message，在code != 0时，展示给前端",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "httpx.JSONResultPaged": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回包code，表明是否正确，在code == 0时，表明服务正常",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "message": {
                    "description": "回报message，在code != 0时，展示给前端",
                    "type": "string",
                    "example": "success"
                },
                "total": {
                    "description": "总数量",
                    "type": "integer"
                }
            }
        },
...
        "list_type.UpdateListTypeReq": {
            "type": "object",
            "required": [
                "description",
                "is_valid"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "description"
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                }
            }
        },
...
        "models.ListType": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "名单类型编码",
                    "type": "string",
                    "example": "telephone__red"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "descriptions"
                },
                "fields": {
                    "description": "这类名单的值被构建的字段",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "telephone",
                        "id_card"
                    ]
                },
                "id": {
                    "description": "主键id",
                    "type": "string",
                    "example": "60d2b17f70d9d2f0db53f866"
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                },
                "namespace": {
                    "description": "命名空间",
                    "type": "string",
                    "example": "anti-fraud"
                }
            }
        },
...
        "models.Namespace": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "命名空间code",
                    "type": "string",
                    "example": "anti-fraud"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "anti fraud use"
                },
                "id": {
                    "description": "主键id",
                    "type": "string",
                    "example": "60d2b17f70d9d2f0db53f866"
                }
            }
        },
        "namespace.CreateNamespaceReq": {
            "type": "object",
            "required": [
                "code",
                "description"
            ],
            "properties": {
                "code": {
                    "description": "命名空间code",
                    "type": "string",
                    "example": "anti-fraud"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "anti fraud use"
                }
            }
        },
        "namespace.UpdateNamespaceReq": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                }
            }
        }
    }
}